% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCatchment.R
\name{addCatchment}
\alias{addCatchment}
\title{Adds a catchment identifier to the sites table}
\usage{
addCatchment(data)
}
\value{
A data.table with site information including catchment id
}
\description{
Adds a catchment identifier to the sites table
}
