% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoCatch.R
\name{addNoCatch}
\alias{addNoCatch}
\title{Pads out count data by run to include passes where zero individuals of a species were caught}
\usage{
addNoCatch(data, spBySite = T, byAge = F)
}
\arguments{
\item{data}{data.table to be expanded. Must include SiteVisit_ID, SiteID, sDate, SPECIES, and RUN as columns}

\item{spBySite}{Logical. If TRUE, expand only to species that were observed at least once at a site within the supplied data. Otherwise, each species will be fully expanded at each site.}

\item{byAge}{Logical. If TRUE, pad out with an AGE==0 and AGE==1 row for each combination of other values.}
}
\value{
A data.table expanded to include site visit/pass/species combos where no individuals were caught
}
\description{
Pads out count data by run to include passes where zero individuals of a species were caught
}
