% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqConnector.R
\name{aqConnector}
\alias{aqConnector}
\title{Connect to Shen aquatics database}
\usage{
aqConnector()
}
\value{
returns \code{con}: An RODBC link to the database for use with the function \code{sqlQuery}
}
\description{
Creates a link to the SHEN MS Access database for import into R.
}
