% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRoot.R
\name{defineRoot}
\alias{defineRoot}
\title{defineRoot}
\usage{
defineRoot()
}
\value{
path to the shenAquatics directory (which contains data, etc.)
}
\description{
Defines the shenAquatics directory path.
}
