% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{discharge}
\alias{discharge}
\title{Converts individual depth and velocity measurements into discharge data}
\usage{
discharge(location, depth, velocity)
}
\arguments{
\item{location}{The distance from the origin for the depth/velocity measurements in meters}

\item{depth}{Depth in meters}

\item{velocity}{Water velocity in m/s}
}
\value{
Discharge in m^3/s
}
\description{
Converts individual depth and velocity measurements into discharge data
}
