% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpac.R
\name{dpac}
\alias{dpac}
\title{Distribute parents among children}
\usage{
dpac(x, countCol = "INSECT_COU", minLevel = "genus", keepCodeCols = NULL)
}
\arguments{
\item{x}{A data.table containing the taxonomic hierarchy (i.e., all taxonomic levels for each taxon) and numbers of individuals}

\item{countCol}{Name of the column containing the counts of individuals}

\item{minLevel}{The taxonomic level at which distribution should be terminated and taxa should be lumped (defaults to genus)}

\item{keepCodeCols}{A vector of names for columns containing taxa codes to be retained in the output}
}
\value{
A data.table with individuals maximally distributed among lower taxonomic levels
}
\description{
Assign higher taxonomic levels to lower
ones based on the relative proportion of lower ones within a sample. The
function is designed for processing of a single sample (or combined samples from a site visit).
See examples for quick operation on multiple samples
}
\examples{
ai<-ai[,dpac(.SD,keepCodeCols=c("TAXON_CODE","TSN")),.(SiteID,sDate)]

}
