% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialSites.R
\name{getSpatialSites}
\alias{getSpatialSites}
\title{Get site data from database and make it a sf object for plotting}
\usage{
getSpatialSites()
}
\description{
Get site data from database and make it a sf object for plotting
}
\details{
Simply a quicker way to get site data for plotting
}
