% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwEst.R
\name{gwEst}
\alias{gwEst}
\title{Estimates deep and shallow groundwater influence from paired air and water temperatures after Hare et al. 2020}
\usage{
gwEst(data, plotFile = NULL, airThresh = 0.5)
}
\arguments{
\item{data}{A data.table that includes the variables airTemp, waterTemp, and date}

\item{plotFile}{An optional file for writing output plots. Plots are shown in the current device in this is left as null.}

\item{airThresh}{A threshold below which air temperatures could be excluded. Hare et al. 2020 says they excluded <1C, but their code suggests it was <0.5C. Adjusting this may strongly affect the amplitude ratio, but is not likely to influence lag much.}
}
\value{
An amplitude ratio (water:air) and lag in days (positive means water is behind air)
}
\description{
Estimates deep and shallow groundwater influence from paired air and water temperatures after Hare et al. 2020
}
\references{
Hare, DK, AM Helton, ZC Johnson, JW Lane, and MA Briggs. 2021. Continental scale analysis of shallow and deep groundwater contributions to streams. Nature Communications 12: 1450.
}
