% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDaymetCov.R
\name{makeDaymetCov}
\alias{makeDaymetCov}
\title{Creates covariates after Kanno et al using daymet data downloaded with downloadDaymet}
\usage{
makeDaymetCov(
  siteId,
  file = paste0("C:/Users/echildress/OneDrive - DOI/Documents/mapShen/data/daymet/",
    siteId, ".csv")
)
}
\arguments{
\item{siteId}{A site ID in the sites table from the database that has lat/long data associated with it}

\item{file}{A file location to save the downloaded data}
}
\value{
A data.table of summarized temp and precip data
}
\description{
Creates covariates after Kanno et al using daymet data downloaded with downloadDaymet
}
