% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFish.R
\name{makeFish}
\alias{makeFish}
\title{Create a dataset with N and detection/non-detection for all species for all site visits in the SHEN fish database}
\usage{
makeFish(nRuns = 4)
}
\value{
returns the data.table
}
\description{
Creates a data.table with detection and number caught for each site visit in the fish database
}
