% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetrics.R
\name{makeMetrics}
\alias{makeMetrics}
\title{Make GLIMPSS and VA SCI indices from a data.table of aquatic macroinvertebrate counts}
\usage{
makeMetrics(
  data,
  byCols = c("SiteID", "sDate"),
  subSize = 200,
  includeValues = F,
  includeScores = F
)
}
\arguments{
\item{data}{A data.table containing counts and required columns listed above}

\item{byCols}{Vector of names of columns to group by for calculations (e.g., site, date)}

\item{subSize}{Size of subsample to use as an integer}

\item{includeValues}{Logical indicating whether to include the columns of intermediate calculations (e.g., percent Ephemeroptera)}

\item{includeScores}{Logical indicating whether to include columns of scores for each index (e.g., SCI score for percent scrapers)}
}
\value{
A data.table of the macroinvertebrate indices by the byCol columns
}
\description{
Calculates GLIMPSS and vA SCI metrics using genus level data (rolled up to family for SCI)
organized in a data.table containing columns: count, order, family, ffg, habit, tolVal,
famTolVal, famFfg. The calculations are made after grouping by the byCol columns.
}
