% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAqSites.R
\name{mapAqSites}
\alias{mapAqSites}
\title{Create a map of aquatic monitoring sites within SHEN}
\usage{
mapAqSites(
  stream = NULL,
  watershed = NULL,
  siteId = NULL,
  fishType = NULL,
  aqinType = NULL,
  text = F,
  add = T,
  pch = 19,
  ...
)
}
\arguments{
\item{stream}{A single value or vector containing the quoted names of streams to be mapped. NULL returns all sites that meet other criteria.}

\item{watershed}{A single value or vector containing the quoted names of watersheds to be mapped. NULL returns all sites that meet other criteria.}

\item{siteId}{A single value or vector containing the quoted names of siteIDs to be mapped. NULL returns all sites that meet other criteria.}

\item{fishType}{A vector of site types for fish for which sites should be returned.}

\item{aqinType}{A vector of site types for aquatic invertebrates for which sites should be returned.}

\item{text}{Logical indicating whether site names should be added next to points.}

\item{add}{Logical indicating whether to add to an existing map (TRUE) or create a new map (FALSE).}
}
\value{
A map with desired sites
}
\description{
Create a map of aquatic monitoring sites within SHEN
}
\details{
This function plots the location of sampling sites for the SHEN aquatic resource monitoring program.
}
