% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapShenStreams.R
\name{mapShenStreams}
\alias{mapShenStreams}
\title{Create a map of streams within and surrounding SHEN}
\usage{
mapShenStreams(
  stream = NULL,
  lwdBoundary = 2,
  wholePark = F,
  add = F,
  streamCol = "blue",
  border = "forestgreen",
  newBox = NULL,
  ...
)
}
\arguments{
\item{stream}{A single value or vector containing the quoted names of streams to be mapped.}

\item{lwdBoundary}{line width for Shen boundary}

\item{wholePark}{Logical indicating whether to map the whole park even if a subset of streams is selected. If wholePark is FALSE, the extent of the map will be trimmed to the area immediately surrounding selected streams.}

\item{add}{Logical indicating whether this should be added to an existing map (TRUE) or if a new map should be created (FALSE)}

\item{border}{color for the SHEN border.}

\item{newBox}{A bbox that defines the area to be plotted.}
}
\value{
A map with desired streams and sites
}
\description{
Create a map of streams within and surrounding SHEN
}
\details{
This function is a wrapper that creates maps of streams within and surrounding SHEN.
}
