% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeOfYear.R
\name{timeOfYear}
\alias{timeOfYear}
\title{similar to yday but converts POSIXct objects as well to a day (or other unit) of the year with portions of the day (or unit) as decimals.}
\usage{
timeOfYear(times, unit = "days", origin = "01-01", numeric = T)
}
\arguments{
\item{times}{A vector of times or dates to be converted}

\item{unit}{Defaults to "days" but can be any other unit accepted by \code{difftime}}

\item{origin}{The starting date for the time of year in the format mm-dd. Defaults to 01-01}
}
\value{
returns numeric values representing progress through the year (i.e. 0-366 days or 0-1 years or 0-8784 hours)
}
\description{
Creates a link to the SHEN MS Access database for import into R.
}
