% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterYear.R
\name{waterYear}
\alias{waterYear}
\title{Indentifies the water year for dates}
\usage{
waterYear(times, origin = "10-01")
}
\arguments{
\item{times}{A vector of times or dates to be converted}

\item{origin}{The starting date for the water year in the format mm-dd. Defaults to 10-01}
}
\value{
returns a numeric value that is the water year for a given date
}
\description{
Water year is typically defined as starting on October 1, but the origin can be defined.
}
