% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDischarge.R
\name{makeDischarge}
\alias{makeDischarge}
\title{Returns calculated discharge by site and date}
\usage{
makeDischarge(excludeSticksBuckets = T)
}
\value{
Discharge in m^3/s
}
\description{
Returns calculated discharge by site and date
}
